Created function to create workspace with defaults
Updated createUser to also create a default workspacefor the user
